+++ src/BaseQuery.js
@@ -5,10 +5,16 @@ const _isInteger = require('lodash/isInteger')
 
 class BaseQuery {
   constructor() {
+    this._initQueryParams()
+  }
+
+  _initQueryParams() {
     this._queryObject = {}
     this._limit = 20
     this._offset = 0
     this._orderBy = null
+    this._keys = null
+    this._expand = null
   }
 
   setQuery(queryObject) {
@@ -45,12 +51,37 @@ class BaseQuery {
     return this
   }
 
+  select(args) {
+    if (args instanceof Array) {
+      this._keys = args.join(',')
+    } else {
+      this._keys = args
+    }
+    return this
+  }
+
+  expand(args) {
+    if (args instanceof Array) {
+      this._expand = args.join(',')
+    } else {
+      this._expand = args
+    }
+    return this
+  }
+
   _handleAllQueryConditions() {
     let conditions = {}
     conditions.limit = this._limit
     conditions.offset = this._offset
     if (this._orderBy) {
       conditions.order_by = this._orderBy
+    }
+    if (this._keys) {
+      conditions.keys = this._keys
+    }
+
+    if (this._expand) {
+      conditions.expand = this._expand
     }
     conditions.where = JSON.stringify(this._queryObject)
     return conditions