+++ src/index.js
@@ -8,6 +8,7 @@ BaaS.FileCategory = require('./FileCategory')
 BaaS.GeoPoint = require('./GeoPoint')
 BaaS.GeoPolygon = require('./GeoPolygon')
 BaaS.getWXACode = require('./getWXACode')
+BaaS.invokeFunction = require('./invokeFunction')
 BaaS.login = require('./auth').login
 BaaS.logout = require('./auth').logout
 BaaS.order = require('./order')