+++ test/BaseQuery.js
@@ -77,15 +77,37 @@ describe('BaseQuery', () => {
     expect(baseQuery._orderBy).to.equal('-amount,price')
   })
 
+  it('#select one', () => {
+    baseQuery.select('-amount')
+    expect(baseQuery._keys).to.equal('-amount')
+  })
+
+  it('#select more', () => {
+    baseQuery.select(['amount', '-price'])
+    expect(baseQuery._keys).to.equal('amount,-price')
+  })
+
+  it('#expand', () => {
+    baseQuery.expand('created_by')
+    expect(baseQuery._expand).to.equal('created_by')
+  })
+
+  it('#expand array args', () => {
+    baseQuery.expand(['created_by', 'test'])
+    expect(baseQuery._expand).to.equal('created_by,test')
+  })
+
   it('#_handleAllQueryConditions', () => {
     let query = new Query()
     query.in('price', randomArray)
     baseQuery.setQuery(query)
     baseQuery.orderBy('-amount')
+    baseQuery.select(['-amount', 'price'])
     expect(baseQuery._handleAllQueryConditions()).to.deep.equal({
       limit: 20,
       offset: 0,
       order_by: '-amount',
+      keys: '-amount,price',
       where: `{"$and":[{"price":{"$in":[${randomArray.join(',')}]}}]}`
     })
   })